feat(agent): rename keymanagement to walletapi
