+++ prism-agent/service/build.sbt
@@ -21,13 +21,13 @@ val commonSettings = Seq(
 lazy val root = project
   .in(file("."))
   .settings(commonSettings)
-  .aggregate(`key-management`, server)
+  .aggregate(`wallet-api`, server)
 
-lazy val `key-management` = project
-  .in(file("key-management"))
+lazy val `wallet-api` = project
+  .in(file("wallet-api"))
   .settings(commonSettings)
   .settings(
-    name := "prism-agent-key-management",
+    name := "prism-agent-wallet-api",
     libraryDependencies ++= keyManagementDependencies
   )
 
@@ -52,4 +52,4 @@ lazy val server = project
     dockerBaseImage := "openjdk:11"
   )
   .enablePlugins(OpenApiGeneratorPlugin, JavaAppPackaging, DockerPlugin)
-  .dependsOn(`key-management`)
+  .dependsOn(`wallet-api`)
