--- prism-agent/service/key-management/src/main/scala/io/iohk/atala/agent/keymanagement/model/CommitmentPurpose.scala
@@ -1,6 +1,0 @@
-package io.iohk.atala.agent.keymanagement.model
-
-enum CommitmentPurpose {
-  case Update extends CommitmentPurpose
-  case Recovery extends CommitmentPurpose
-}
