--- prism-agent/service/key-management/src/main/scala/io/iohk/atala/agent/keymanagement/model/ManagedDIDTemplate.scala
@@ -1,14 +1,0 @@
-package io.iohk.atala.agent.keymanagement.model
-
-import io.iohk.atala.castor.core.model.did.{Service, VerificationRelationship}
-
-final case class ManagedDIDTemplate(
-    storage: String,
-    publicKeys: Seq[DIDPublicKeyTemplate],
-    services: Seq[Service]
-)
-
-final case class DIDPublicKeyTemplate(
-    id: String,
-    purpose: VerificationRelationship
-)
