--- prism-agent/service/key-management/src/main/scala/io/iohk/atala/agent/keymanagement/model/error/CreateManagedDIDError.scala
@@ -1,13 +1,0 @@
-package io.iohk.atala.agent.keymanagement.model.error
-
-import io.iohk.atala.castor.core.model.did.PrismDID
-import io.iohk.atala.castor.core.model.error.DIDOperationError
-
-sealed trait CreateManagedDIDError
-
-object CreateManagedDIDError {
-  final case class DIDAlreadyExists(did: PrismDID) extends CreateManagedDIDError
-  final case class KeyGenerationError(cause: Throwable) extends CreateManagedDIDError
-  final case class WalletStorageError(cause: Throwable) extends CreateManagedDIDError
-  final case class OperationError(cause: DIDOperationError) extends CreateManagedDIDError
-}
