--- prism-agent/service/key-management/src/main/scala/io/iohk/atala/agent/keymanagement/storage/DIDNonSecretStorage.scala
@@ -1,14 +1,0 @@
-package io.iohk.atala.agent.keymanagement.storage
-
-import io.iohk.atala.castor.core.model.did.{PrismDID, PublishedDIDOperation}
-import zio.*
-
-private[keymanagement] trait DIDNonSecretStorage {
-
-  def getCreatedDID(did: PrismDID): Task[Option[PublishedDIDOperation.Create]]
-
-  def saveCreatedDID(did: PrismDID, createOp: PublishedDIDOperation.Create): Task[Unit]
-
-  def listCreatedDID: Task[Seq[PrismDID]]
-
-}
