+++ prism-agent/service/wallet-api/src/main/scala/io/iohk/atala/agent/walletapi/model/CommitmentPurpose.scala
@@ -1,0 +1,6 @@
+package io.iohk.atala.agent.walletapi.model
+
+enum CommitmentPurpose {
+  case Update extends CommitmentPurpose
+  case Recovery extends CommitmentPurpose
+}
