+++ prism-agent/service/wallet-api/src/main/scala/io/iohk/atala/agent/walletapi/model/ManagedDIDTemplate.scala
@@ -1,0 +1,14 @@
+package io.iohk.atala.agent.walletapi.model
+
+import io.iohk.atala.castor.core.model.did.{Service, VerificationRelationship}
+
+final case class ManagedDIDTemplate(
+    storage: String,
+    publicKeys: Seq[DIDPublicKeyTemplate],
+    services: Seq[Service]
+)
+
+final case class DIDPublicKeyTemplate(
+    id: String,
+    purpose: VerificationRelationship
+)
