+++ prism-agent/service/wallet-api/src/main/scala/io/iohk/atala/agent/walletapi/util/Prism14CompatUtil.scala
@@ -1,0 +1,16 @@
+package io.iohk.atala.agent.walletapi.util
+
+import com.ionspin.kotlin.bignum.integer.BigInteger
+import com.ionspin.kotlin.bignum.integer.util.ConversionUtilsKt
+
+object Prism14CompatUtil {
+
+  extension (n: BigInteger) {
+    def toScalaBigInt: BigInt = BigInt(ConversionUtilsKt.toTwosComplementByteArray(n))
+  }
+
+  extension (n: BigInt) {
+    def toKotlinBigInt: BigInteger = ConversionUtilsKt.fromTwosComplementByteArray(BigInteger.Companion, n.toByteArray)
+  }
+
+}
